+++ flux-common-aws/src/main/java/com/danielgmyers/flux/util/AwsRetryUtils.java
@@ -14,18 +14,13 @@
  *   limitations under the License.
  */
 
-package com.danielgmyers.flux.clients.swf.util;
+package com.danielgmyers.flux.util;
 
-import java.lang.reflect.Method;
 import java.time.Duration;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.function.Supplier;
 
-import com.danielgmyers.flux.clients.swf.FluxCapacitorImpl;
-import com.danielgmyers.flux.step.StepApply;
-import com.danielgmyers.flux.step.WorkflowStep;
-import com.danielgmyers.flux.step.internal.WorkflowStepUtil;
 import com.danielgmyers.metrics.MetricRecorder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,14 +31,16 @@ import software.amazon.awssdk.core.exception.SdkServiceException;
 /**
  * Utilities for retrying operations and calculating backoff times.
  */
-public final class RetryUtils {
+public final class AwsRetryUtils {
 
-    private static final Logger log = LoggerFactory.getLogger(RetryUtils.class);
+    private static final Logger log = LoggerFactory.getLogger(AwsRetryUtils.class);
 
-    private RetryUtils() {}
+    private AwsRetryUtils() {}
 
     private static final String THROTTLING_MESSAGE = "rate exceeded";
     private static final Set<Class<?>> RETRYABLE_INNER_CLIENT_EXCEPTIONS = new HashSet<>();
+
+    private static final double INLINE_EXPONENTIAL_BACKOFF_RATE = 1.25;
 
     static {
         RETRYABLE_INNER_CLIENT_EXCEPTIONS.add(java.net.SocketException.class);
@@ -63,6 +60,7 @@ public final class RetryUtils {
      * Retries a given function up to maxAttempts times, with exponential backoff and jitter, a minimum
      * retry delay of minRetryDelay, and a maximum retry delay of maxRetryDelay.
      * It records time and various count metrics using the specified operation prefix.
+     * Specifically handles throttling exceptions from AWS services and retryable client SDK exceptions.
      */
     public static <T> T executeWithInlineBackoff(Supplier<T> function, long maxAttempts, Duration minRetryDelay,
                                                  Duration maxRetryDelay, MetricRecorder metrics, String operationPrefix) {
@@ -92,7 +90,7 @@ public final class RetryUtils {
                 metrics.endDuration(timeMetricName);
 
                 if (!e.isThrottlingException()) {
-                    // SWF doesn't (always?) use status code 429 for throttling, so we have to check the message too :(
+                    // Some services don't use status code 429 for throttling, so we have to check the message too :(
                     String message = e.getMessage();
                     if (message == null || !message.toLowerCase().contains(THROTTLING_MESSAGE)) {
                         metrics.addCount(failureCountMetricName, 1);
@@ -109,9 +107,9 @@ public final class RetryUtils {
 
                 // we'll always start backing off after the second retry, up to maxRetryDelay per retry, with 10% jitter.
                 // we'll treat this as milliseconds and sleep for this long.
-                long sleepTimeMillis = calculateRetryBackoff(1, minRetryDelay.toMillis(), maxRetryDelay.toMillis(), retry, 10,
-                    FluxCapacitorImpl.DEFAULT_EXPONENTIAL_BACKOFF_BASE);
-                log.debug("Throttled by SWF, retrying after {}ms.", sleepTimeMillis);
+                long sleepTimeMillis = RetryUtils.calculateRetryBackoff(1, minRetryDelay.toMillis(), maxRetryDelay.toMillis(),
+                        retry, 10, INLINE_EXPONENTIAL_BACKOFF_RATE);
+                log.debug("Throttled by service, retrying after {}ms.", sleepTimeMillis);
                 try {
                     Thread.sleep(sleepTimeMillis);
                 } catch (InterruptedException ie) {
@@ -124,7 +122,6 @@ public final class RetryUtils {
                 }
                 metrics.endDuration(timeMetricName);
 
-
                 metrics.addCount(retryableClientExceptionCountMetricName, 1);
                 retry++;
                 if (retry >= maxAttempts) {
@@ -136,8 +133,8 @@ public final class RetryUtils {
 
                 // we'll always start backing off after the second retry, up to maxRetryDelay per retry, with 10% jitter.
                 // we'll treat this as milliseconds and sleep for this long.
-                long sleepTimeMillis = calculateRetryBackoff(1, minRetryDelay.toMillis(), maxRetryDelay.toMillis(), retry, 10,
-                    FluxCapacitorImpl.DEFAULT_EXPONENTIAL_BACKOFF_BASE);
+                long sleepTimeMillis = RetryUtils.calculateRetryBackoff(1, minRetryDelay.toMillis(), maxRetryDelay.toMillis(),
+                        retry, 10, INLINE_EXPONENTIAL_BACKOFF_RATE);
                 log.debug("Retryable client Exception, retrying after {}ms.", sleepTimeMillis);
                 try {
                     Thread.sleep(sleepTimeMillis);
@@ -146,47 +143,6 @@ public final class RetryUtils {
                 }
             }
         }
-    }
-
-    /**
-     * Calculates a retry's backoff time in seconds, respecting the specified particular workflow step's
-     * configured retry settings.
-     */
-    public static long calculateRetryBackoffInSeconds(WorkflowStep stepToRetry, long retryAttempt,
-                                                      double exponentialBackoffBase) {
-        Method applyMethod = WorkflowStepUtil.getUniqueAnnotatedMethod(stepToRetry.getClass(), StepApply.class);
-        StepApply applyConfig = applyMethod.getAnnotation(StepApply.class);
-
-        double backoffBase = exponentialBackoffBase;
-        if (applyConfig.exponentialBackoffBase() > 0.0) {
-            backoffBase = applyConfig.exponentialBackoffBase();
-        }
-
-        return calculateRetryBackoff(applyConfig.retriesBeforeBackoff(), applyConfig.initialRetryDelaySeconds(),
-                                     applyConfig.maxRetryDelaySeconds(), retryAttempt, applyConfig.jitterPercent(),
-                                     backoffBase);
-    }
-
-    private static long calculateRetryBackoff(long retriesBeforeBackoff, long initialRetryDelay, long maxRetryDelay,
-                                             long retryAttempt, long jitterPercent, double exponentialBackoffBase) {
-        // We subtract from the retryAttempt number so that some number of retries use the initial retry time.
-        // At a minimum, we want to subtract 1 so that we don't skip the first retry using the initial retry time.
-        // We can adjust this to e.g. retryAttempt - 3 so that the first 3 retry attempts use the initial retry time.
-        long power = Math.max(0, retryAttempt - retriesBeforeBackoff);
-
-        long retryDelay = (long)(initialRetryDelay * Math.pow(exponentialBackoffBase, power));
-
-        // make sure we don't go over the pre-jitter max delay
-        retryDelay = Math.min(retryDelay, maxRetryDelay);
-
-        // calculate the max jitter duration
-        final double jitterMultiplier = ((double)jitterPercent / 100.0);
-        final long maxJitter = (jitterPercent == 0 ? 0 : (long)(retryDelay * jitterMultiplier));
-
-        // First we reduce retryDelay by maxJitter, then add a random integer in [0, 2 * maxJitter] to the retry time.
-        retryDelay = (retryDelay - maxJitter + (long)Math.rint(2 * Math.random() * maxJitter));
-
-        return retryDelay;
     }
 
     private static Throwable getInnerRetryableClientException(SdkClientException ex) {
